+++ src/test/java/com/abab/lectureRegister/service/RegistrationTest.java
@@ -1,10 +1,9 @@
-package com.abab.lectureRegister.registration;
+package com.abab.lectureRegister.service;
 
 import com.abab.lectureRegister.exception.custom.LectureAlreadyRegisteredException;
 import com.abab.lectureRegister.model.Lecture;
 import com.abab.lectureRegister.model.Registration;
 import com.abab.lectureRegister.repository.RegistrationRepository;
-import com.abab.lectureRegister.service.RegistrationService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
